fix: 统一 Python 版本为 3.9.11 + 完善 Windows 文档
- 更新 requirements.txt 为 Python 3.9.11 兼容版本
- 创建 WINDOWS_INSTALL.md 详细安装指南
- 修复 init_db.py 兼容性问题（显式指定 ID）
- 添加常见问题解决方案

Windows 用户请确保：
1. 安装 Python 3.9.11（必须）
2. 使用虚拟环境
3. 使用 -m 参数运行 init_db

+++ backend/app/models/config.py
@@ -10,7 +10,7 @@ class SystemConfig(Base):
     
     __tablename__ = "system_configs"
     
-    id = Column(BigInteger, primary_key=True, autoincrement=True, comment="主键 ID")
+    id = Column(Integer, primary_key=True, autoincrement=True, comment="主键 ID")
     config_key = Column(String(100), nullable=False, unique=True, comment="配置键（如：claude_api_key）")
     config_value = Column(Text, nullable=False, comment="配置值（加密存储）")
     config_type = Column(String(20), nullable=False, default="string", comment="配置类型：string/number/boolean/json")
